rename resource to httpadapter
