+++ backend/resource/httpadapter/doc.go
@@ -1,0 +1,2 @@
+// Package httpadapter provides support for handling resource calls using an http.Handler.
+package httpadapter
