+++ backend/resource/httpadapter/handler_test.go
@@ -1,4 +1,4 @@
-package resource
+package httpadapter
 
 import (
 	"context"
