+++ backend/resource/httpadapter/response_writer.go
@@ -1,4 +1,4 @@
-package resource
+package httpadapter
 
 import (
 	"bytes"
